Add 'avatar_img' column to User table
This avatar image is provided by Reddit when hitting the GET /api/v1/me
endpoint. Used in the Header component in the web app.

+++ src/database/repositories/auth.js
@@ -1,36 +1,32 @@
 const db = require('../config')
 
-const readUser = async userId => {
+const readUser = async (userId) => {
   return await db('user')
     .columns(
       process.env.NODE_ENV === 'production'
-        ? ['id', 'username', 'email', 'frequency']
+        ? ['id', 'username', 'avatar_img']
         : '*',
     )
     .where({ id: userId })
     .first()
 }
 
-const createUser = async data => {
+const createUser = async (data) => {
   await db('user')
     .insert(data)
     .returning(
       process.env.NODE_ENV === 'production'
-        ? ['id', 'username', 'email', 'frequency']
+        ? ['id', 'username', 'avatar_img']
         : '*',
     )
 }
 
 const updateUser = async (userId, updates) => {
-  await db('user')
-    .where({ id: userId })
-    .update(updates)
+  await db('user').where({ id: userId }).update(updates)
 }
 
-const removeUser = async userId => {
-  await db('user')
-    .where({ id: userId })
-    .delete()
+const removeUser = async (userId) => {
+  await db('user').where({ id: userId }).delete()
 }
 
 module.exports = {
